@@ -2,7 +2,7 @@ package ast
 
 func indent(lv int) string {
 	var str string
-	for i := 0; i < lv; i++ {
+	for range lv {
 		str += "  "
 	}
 	return str
