@@ -58,7 +58,7 @@ func BenchmarkFormatter(b *testing.B) {
 		return
 	}
 
-	for i := 0; i < b.N; i++ {
+	for b.Loop() {
 		New(c).Format(vcl)
 	}
 }
