@@ -63,7 +63,7 @@ func Querystring_get(ctx *context.Context, args ...value.Value) (value.Value, er
 	// ?name  => should return empty string, but returns empty string
 	// ?name= => should return not set, but returns empty string
 	// so we try to parse from RawQuery string, not using url.Value
-	for _, query := range strings.Split(qs, "&") {
+	for query := range strings.SplitSeq(qs, "&") {
 		sp := strings.SplitN(query, "=", 2)
 		if len(sp) < 2 || sp[0] == "" {
 			continue
